+++ src/Diagnostics/Diagnostics.jl
@@ -61,5 +61,6 @@ include("vortex_impulse.jl")
 include("stretching.jl")
 include("spectra.jl")
 include("spectra_helicity.jl")
+include("integral_scale.jl")
 
 end