+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add `Diagnostics.integral_lengthscale`.
+
 ## [0.28.3] - 2025-06-23
 
 ### Fixed